+++ lib/dom_tools.dart
@@ -5,6 +5,7 @@ export 'src/dom_tools_base.dart';
 export 'src/dom_tools_js.dart';
 export 'src/dom_tools_css.dart';
 export 'src/dom_tools_document.dart';
+export 'src/dom_tools_dialog.dart';
 export 'src/dom_tools_track.dart';
 export 'src/dom_tools_paint.dart';
 export 'src/perspective_filter.dart';