+++ test/dom_tools_test.dart
@@ -1,6 +1,8 @@
 import 'package:dom_tools/dom_tools.dart';
 import 'package:test/test.dart';
 
+@TestOn('browser')
+
 void main() {
   group('A group of tests', () {
     setUp(() {